feat: Configuration Change 대응을 위해 CanvasHistory 싱글톤으로 구현
